+++ src/OrchardCore.Modules/OrchardCore.Users/Abstractions/IControllerAbstraction.cs
@@ -1,0 +1,7 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace OrchardCore.Users.Abstractions;
+public interface IControllerAbstraction<T> where T : Controller
+{
+    T BindController { get; set; }
+}
